PORTALS-2628 apps: - use vitest, reducing configuration overhead and allowing us to use esm more easily than jest would allow - Remove babel and jest
synapse-react-client:
- break circular dependencies in lib/utils

vite-config:
- add and export reusable vitest-config that were preventing tests from breaking
