+++ apps/SageAccountWeb/package.json
@@ -27,13 +27,8 @@
     "universal-cookie": "^4.0.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.0",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^27.5.0",
     "@types/katex": "^0.5.0",
     "@types/node": "11.10.4",
     "@types/plotly.js": "^1.54.10",
@@ -43,13 +38,12 @@
     "@types/react-plotly.js": "^2.6.0",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-tooltip": "^4.2.4",
+    "@vitest/coverage-c8": "^0.30.1",
+    "@vitest/ui": "^0.30.1",
     "assert": "^2.0.0",
     "buffer": "^6.0.3",
     "https-browserify": "^1.0.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jest-html-reporter": "^3.7.0",
     "jsdom": "^21.1.1",
     "katex": "^0.10.0-rc.1",
     "markdown-it": "^13.0.1",
@@ -80,13 +74,15 @@
     "util": "^0.12.4",
     "vite": "^4.3.4",
     "vite-config": "workspace:*",
+    "vitest": "^0.30.1",
     "whatwg-fetch": "^3.6.2"
   },
   "scripts": {
     "clean": "rimraf build coverage",
     "start": "vite",
     "build": "vite build",
-    "test": "jest --watchAll=false"
+    "test": "vitest",
+    "test:coverage": "vitest run --coverage"
   },
   "browserslist": {
     "production": [
