+++ apps/SageAccountWeb/src/tests/AppInitialization.test.tsx
@@ -4,6 +4,7 @@ import LoginPage from '../LoginPage'
 import React from 'react'
 import { MemoryRouter } from 'react-router-dom'
 import FullContextProvider from 'synapse-react-client/dist/utils/FullContextProvider'
+import { describe, it } from 'vitest'
 
 describe('App initialization tests', () => {
   it.skip('renders a branded experience', async () => {
