+++ apps/SageAccountWeb/vite.config.ts
@@ -1,4 +1,12 @@
-import { defineConfig } from 'vite'
-import sharedConfig from 'vite-config'
+import { vitestConfig } from 'vite-config'
+import { mergeConfig, defineConfig } from 'vitest/config'
 
-export default defineConfig(sharedConfig)
+export default mergeConfig(
+  vitestConfig,
+  defineConfig({
+    test: {
+      include: ['src/tests/**/*.test.[jt]s?(x)'],
+      setupFiles: ['src/tests/setupTests.ts'],
+    },
+  }),
+)
