+++ apps/portals/package.json
@@ -36,17 +36,14 @@
     "start": "vite",
     "build": "pnpm copy-assets; pnpm save-build-date; vite build",
     "copy-test-configuration": "cp -r src/test-configuration/ src/config/",
-    "test": "pnpm run copy-test-configuration && jest test --watchAll=false",
+    "test": "pnpm run copy-test-configuration && vitest",
+    "test:coverage": "pnpm run copy-test-configuration && vitest run --coverage",
     "generate-sitemap": "node sitemap/generate-sitemap.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.0",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^27.5.0",
     "@types/katex": "^0.5.0",
     "@types/lodash": "^4.14.191",
     "@types/node": "11.10.4",
@@ -55,15 +52,12 @@
     "@types/react-dom": "^18.0.6",
     "@types/react-plotly.js": "^2.6.0",
     "@types/react-router-dom": "^5.3.3",
+    "@vitest/coverage-c8": "^0.30.1",
+    "@vitest/ui": "^0.30.1",
     "assert": "^2.0.0",
-    "babel-jest": "^27.5.1",
-    "babel-plugin-transform-vite-meta-env": "^1.0.3",
     "buffer": "^6.0.3",
     "https-browserify": "^1.0.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jest-html-reporter": "^3.7.0",
     "jsdom": "^21.1.1",
     "katex": "^0.10.0-rc.1",
     "markdown-it": "^13.0.1",
@@ -94,6 +88,7 @@
     "util": "^0.12.4",
     "vite": "^4.3.4",
     "vite-config": "workspace:*",
+    "vitest": "^0.30.1",
     "whatwg-fetch": "^3.6.2"
   },
   "browserslist": {
