+++ apps/portals/src/README.md
@@ -13,7 +13,7 @@ Structure of the codebase
 ├── RouteResolver.tsx     # For any subpage not in Home/Explore this hooks up the configuration to that URL
 ├── config                # The main configuration folder read for the app 
 ├── test-configuration    # Folder with an example config that gets used for testing
-├── index.tsx             # Boilerplate code that hooks the application up with index.html
+├── index.ts             # Boilerplate code that hooks the application up with index.html
 ├── portal-components     # portal specific components that are not related to layout
 ├── configurations        # folder containing all the portal configs
 ├── tests                 # all the tests for the application
