+++ apps/portals/src/tests/RouteControl.test.tsx
@@ -1,12 +1,13 @@
 import * as React from 'react'
 import { RouteControl, RouteControlProps } from '../RouteControl'
 import { render, screen } from '@testing-library/react'
+import { describe, expect, it, vi } from 'vitest'
 
 describe('RouteControl works', () => {
   const customRoutes = ['route1', 'route2']
   const props: RouteControlProps = {
     customRoutes,
-    handleChanges: jest.fn(),
+    handleChanges: vi.fn(),
     isSelected: (route) => route === 'route1',
   }
   it('renders the route tabs', () => {
