+++ apps/portals/src/tests/SynapseComponent.test.tsx
@@ -4,6 +4,7 @@ import { MemoryRouter } from 'react-router-dom'
 import { SynapseComponent } from '../SynapseComponent'
 import { SynapseConfig } from '../types/portal-config'
 import FullContextProvider from 'synapse-react-client/dist/utils/FullContextProvider'
+import { describe, it } from 'vitest'
 
 describe('SynapseComponent tests', () => {
   it('renders SRC components correctly', () => {
