+++ apps/portals/src/tests/portal-components/StatefulButtonControlWrapper.test.tsx
@@ -4,6 +4,7 @@ import { MemoryRouter } from 'react-router-dom'
 import StatefulButtonControlWrapper, {
   StatefulButtonControlWrapperProps,
 } from '../../portal-components/StatefulButtonControlWrapper'
+import { describe, it } from 'vitest'
 
 // Michael TODO: Investigate why SynapseComponents has undefined CardContainer/CardContainerLogic
 // exclusively at the start of this test
