+++ apps/portals/src/tests/portal-components/injectPropsIntoConfig.test.ts
@@ -1,5 +1,6 @@
 import injectPropsIntoConfig from '../../portal-components/injectPropsIntoConfig'
 import { MarkdownSynapseProps } from 'synapse-react-client/dist/containers/markdown/MarkdownSynapse'
+import { describe, expect, it } from 'vitest'
 
 describe('injectPropsIntoConfig works', () => {
   it('works with markdown and injecting the ownerid', () => {
