--- apps/synapse-oauth-signin/babel.config.js
@@ -1,9 +1,0 @@
-// Babel configuration currently only used by Jest.
-module.exports = {
-  presets: [
-    '@babel/preset-env',
-    ['@babel/preset-typescript', { allowDeclareFields: true }],
-    '@babel/preset-react',
-  ],
-  sourceMap: 'inline',
-}
