+++ apps/synapse-oauth-signin/package.json
@@ -27,10 +27,6 @@
     "synapse-react-client": "workspace:*"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.0",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.0",
@@ -42,14 +38,13 @@
     "@types/react-plotly.js": "^2.6.0",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-tooltip": "^4.2.4",
+    "@vitest/coverage-c8": "^0.30.1",
+    "@vitest/ui": "^0.30.1",
     "assert": "^2.0.0",
     "buffer": "^6.0.3",
     "https-browserify": "^1.0.0",
     "identity-obj-proxy": "^3.0.0",
     "isomorphic-fetch": "^2.2.1",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jest-html-reporter": "^3.7.0",
     "jsdom": "^21.1.1",
     "katex": "^0.10.0-rc.1",
     "lint-staged": "^13.1.2",
@@ -83,13 +78,15 @@
     "url": "^0.11.0",
     "util": "^0.12.4",
     "vite": "^4.3.4",
-    "vite-config": "workspace:*"
+    "vite-config": "workspace:*",
+    "vitest": "^0.30.1"
   },
   "scripts": {
     "clean": "rimraf build coverage",
     "start": "vite",
     "build": "vite build",
-    "test": "jest --watchAll=false",
+    "test": "vitest",
+    "test:coverage": "vitest run --coverage",
     "lint": "eslint src"
   },
   "browserslist": {
