+++ apps/synapse-oauth-signin/src/test/setupTests.ts
@@ -1,6 +1,7 @@
-import 'isomorphic-fetch' // polyfill for fetch
-import 'raf/polyfill' // polyfill for requestAnimationFrame
+import fetch from 'isomorphic-fetch' // polyfill for fetch
 import { server } from '../mocks/server'
+import { cleanup } from '@testing-library/react'
+import { afterAll, afterEach, beforeAll, vi } from 'vitest'
 
 declare const global: any
 global.markdownit = require('markdown-it')
@@ -17,6 +18,9 @@ global.markdownitInlineComments = require('markdown-it-inline-comments')
 global.markdownitBr = require('markdown-it-br')
 global.markdownitMath = require('markdown-it-synapse-math')
 
+// Need to add this line
+global.fetch = fetch
+
 // Line below is used because plotly has a dependency on mapbox-gl
 // which requires a browser env and doesn't provide support for headless
 // js testing, so we shim the function below.
@@ -32,6 +36,11 @@ afterEach(() => server.resetHandlers())
 // Clean up after the tests are finished.
 afterAll(() => server.close())
 
+// Clean up after each test case (e.g. clearing jsdom)
+afterEach(() => {
+  cleanup()
+})
+
 // Mock window.location
 // https://www.benmvp.com/blog/mocking-window-location-methods-jest-jsdom/
 const oldWindowLocation = window.location
@@ -44,11 +53,11 @@ beforeAll(() => {
       ...Object.getOwnPropertyDescriptors(oldWindowLocation),
       assign: {
         configurable: true,
-        value: jest.fn(),
+        value: vi.fn(),
       },
       replace: {
         configurable: true,
-        value: jest.fn(),
+        value: vi.fn(),
       },
     },
   )
