+++ packages/synapse-react-client/src/lib/containers/UserCardListRotate.tsx
@@ -12,8 +12,7 @@ import {
 } from '../utils/synapseTypes/Table'
 import { ColumnTypeEnum } from '../utils/synapseTypes'
 import UserCardList from './UserCardList'
-import { UserCardSize } from './UserCard'
-import { LARGE_USER_CARD } from '../utils/SynapseConstants'
+import { LARGE_USER_CARD, UserCardSize } from '../utils/SynapseConstants'
 import { Button } from 'react-bootstrap'
 import { useSynapseContext } from '../utils/SynapseContext'
 import { LoadingUserCardMedium } from './UserCardMedium'
