+++ packages/synapse-react-client/src/lib/utils/SynapseError.ts
@@ -1,0 +1,6 @@
+/**
+ * Error message returned by the Synapse Backend
+ */
+export type SynapseError = {
+  reason: string
+}
