+++ packages/synapse-react-client/src/lib/utils/functions/EntityTypeUtils.ts
@@ -1,35 +1,35 @@
 import {
+  Dataset,
+  DATASET_COLLECTION_CONCRETE_TYPE_VALUE,
+  DATASET_CONCRETE_TYPE_VALUE,
+  DatasetCollection,
   Entity,
+  ENTITY_CONCRETE_TYPE,
+  ENTITY_VIEW_CONCRETE_TYPE_VALUE,
+  ENTITY_VIEW_TYPE_MASK_FILE,
   EntityHeader,
   EntityType,
   EntityView,
-  ENTITY_CONCRETE_TYPE,
-  ENTITY_VIEW_TYPE_MASK_FILE,
-  ProjectHeader,
-  Dataset,
-  Table,
-  TableEntity,
-  SubmissionView,
-  VersionableEntity,
-  VIEW_CONCRETE_TYPE_VALUES,
-  View,
-  TABLE_CONCRETE_TYPE_VALUES,
-  SUBMISSION_VIEW_CONCRETE_TYPE_VALUE,
-  DATASET_CONCRETE_TYPE_VALUE,
-  ENTITY_VIEW_CONCRETE_TYPE_VALUE,
-  DatasetCollection,
-  DATASET_COLLECTION_CONCRETE_TYPE_VALUE,
   FILE_ENTITY_CONCRETE_TYPE_VALUE,
-  LINK_CONCRETE_TYPE_VALUE,
   FOLDER_CONCRETE_TYPE_VALUE,
+  LINK_CONCRETE_TYPE_VALUE,
   PROJECT_CONCRETE_TYPE_VALUE,
+  ProjectHeader,
+  SUBMISSION_VIEW_CONCRETE_TYPE_VALUE,
+  SubmissionView,
+  Table,
+  TABLE_CONCRETE_TYPE_VALUES,
+  TABLE_ENTITY_CONCRETE_TYPE_VALUE,
+  TableEntity,
+  VersionableEntity,
+  View,
+  VIEW_CONCRETE_TYPE_VALUES,
 } from '../synapseTypes'
 import { Hit } from '../synapseTypes/Search'
 import {
-  MaterializedView,
   MATERIALIZED_VIEW_CONCRETE_TYPE_VALUE,
+  MaterializedView,
 } from '../synapseTypes/Table/MaterializedView'
-import { TABLE_ENTITY_CONCRETE_TYPE_VALUE } from '../synapseTypes/Table/TableEntity'
 import { isTypeViaConcreteTypeFactory } from './TypeUtils'
 import { DOCKER_REPOSITORY_CONCRETE_TYPE_VALUE } from '../synapseTypes/docker/DockerRepository'
