+++ packages/synapse-react-client/src/lib/utils/index.ts
