+++ packages/synapse-react-client/src/lib/utils/synapseTypes/SignedToken/InviteeVerificationSignedToken.ts
@@ -1,12 +1,11 @@
 import { SignedTokenInterface } from './SignedTokenInterface'
 import { isTypeViaConcreteTypeFactory } from '../../functions/TypeUtils'
+import {
+  INVITEE_VERIFICATION_CONCRETE_TYPE,
+  INVITEE_VERIFICATION_CONCRETE_TYPE_VALUE,
+} from './SignedTokenConcreteType'
 
 // https://rest-docs.synapse.org/rest/org/sagebionetworks/repo/model/InviteeVerificationSignedToken.html
-export const INVITEE_VERIFICATION_CONCRETE_TYPE_VALUE =
-  'org.sagebionetworks.repo.model.InviteeVerificationSignedToken'
-export type INVITEE_VERIFICATION_CONCRETE_TYPE =
-  typeof INVITEE_VERIFICATION_CONCRETE_TYPE_VALUE
-
 export interface InviteeVerificationSignedToken extends SignedTokenInterface {
   readonly concreteType: INVITEE_VERIFICATION_CONCRETE_TYPE
   readonly inviteeId: string //The ID of the invitee.
