+++ packages/synapse-react-client/src/lib/utils/synapseTypes/SignedToken/JoinTeamSignedToken.ts
@@ -1,11 +1,11 @@
 import { isTypeViaConcreteTypeFactory } from '../../functions/TypeUtils'
 import { SignedTokenInterface } from './SignedTokenInterface'
+import {
+  JOIN_TEAM_CONCRETE_TYPE,
+  JOIN_TEAM_CONCRETE_TYPE_VALUE,
+} from './SignedTokenConcreteType'
 
 // https://rest-docs.synapse.org/rest/org/sagebionetworks/repo/model/JoinTeamSignedToken.html
-export const JOIN_TEAM_CONCRETE_TYPE_VALUE =
-  'org.sagebionetworks.repo.model.JoinTeamSignedToken'
-export type JOIN_TEAM_CONCRETE_TYPE = typeof JOIN_TEAM_CONCRETE_TYPE_VALUE
-
 export interface JoinTeamSignedToken extends SignedTokenInterface {
   readonly concreteType: JOIN_TEAM_CONCRETE_TYPE
   readonly userId: string //The ID of the user who is acting to add the new member to the Team. The HMAC in the token authenticates that the request is being made by this user.
