+++ packages/synapse-react-client/src/lib/utils/synapseTypes/SignedToken/MembershipInvtnSignedToken.ts
@@ -1,12 +1,11 @@
 import { isTypeViaConcreteTypeFactory } from '../../functions/TypeUtils'
 import { SignedTokenInterface } from './SignedTokenInterface'
+import {
+  MEMBERSHIP_INVITATION_CONCRETE_TYPE,
+  MEMBERSHIP_INVITATION_CONCRETE_TYPE_VALUE,
+} from './SignedTokenConcreteType'
 
 // https://rest-docs.synapse.org/rest/org/sagebionetworks/repo/model/MembershipInvtnSignedToken.html
-export const MEMBERSHIP_INVITATION_CONCRETE_TYPE_VALUE =
-  'org.sagebionetworks.repo.model.MembershipInvtnSignedToken'
-export type MEMBERSHIP_INVITATION_CONCRETE_TYPE =
-  typeof MEMBERSHIP_INVITATION_CONCRETE_TYPE_VALUE
-
 export interface MembershipInvtnSignedToken extends SignedTokenInterface {
   readonly concreteType: MEMBERSHIP_INVITATION_CONCRETE_TYPE
   readonly membershipInvitationId: string //The ID of the membership invitation.
