+++ packages/synapse-react-client/src/lib/utils/synapseTypes/SignedToken/SignedTokenConcreteType.ts
@@ -1,0 +1,15 @@
+export const MEMBERSHIP_INVITATION_CONCRETE_TYPE_VALUE =
+  'org.sagebionetworks.repo.model.MembershipInvtnSignedToken'
+export type MEMBERSHIP_INVITATION_CONCRETE_TYPE =
+  typeof MEMBERSHIP_INVITATION_CONCRETE_TYPE_VALUE
+export const JOIN_TEAM_CONCRETE_TYPE_VALUE =
+  'org.sagebionetworks.repo.model.JoinTeamSignedToken'
+export type JOIN_TEAM_CONCRETE_TYPE = typeof JOIN_TEAM_CONCRETE_TYPE_VALUE
+export const INVITEE_VERIFICATION_CONCRETE_TYPE_VALUE =
+  'org.sagebionetworks.repo.model.InviteeVerificationSignedToken'
+export type INVITEE_VERIFICATION_CONCRETE_TYPE =
+  typeof INVITEE_VERIFICATION_CONCRETE_TYPE_VALUE
+export type SIGNED_TOKEN_CONCRETE_TYPE =
+  | JOIN_TEAM_CONCRETE_TYPE
+  | INVITEE_VERIFICATION_CONCRETE_TYPE
+  | MEMBERSHIP_INVITATION_CONCRETE_TYPE
