+++ packages/synapse-react-client/src/lib/utils/synapseTypes/Table/ConcreteType.ts
@@ -1,0 +1,28 @@
+import { MATERIALIZED_VIEW_CONCRETE_TYPE_VALUE } from './MaterializedView'
+import { ENTITY_VIEW_CONCRETE_TYPE_VALUE } from './EntityView'
+import { SUBMISSION_VIEW_CONCRETE_TYPE_VALUE } from './SubmissionView'
+import {
+  DATASET_COLLECTION_CONCRETE_TYPE_VALUE,
+  DATASET_CONCRETE_TYPE_VALUE,
+} from './Dataset'
+
+export const TABLE_ENTITY_CONCRETE_TYPE_VALUE =
+  'org.sagebionetworks.repo.model.table.TableEntity'
+export type TABLE_ENTITY_CONCRETE_TYPE = typeof TABLE_ENTITY_CONCRETE_TYPE_VALUE
+
+export const VIEW_CONCRETE_TYPE_VALUES = [
+  ENTITY_VIEW_CONCRETE_TYPE_VALUE,
+  SUBMISSION_VIEW_CONCRETE_TYPE_VALUE,
+  DATASET_CONCRETE_TYPE_VALUE,
+  DATASET_COLLECTION_CONCRETE_TYPE_VALUE,
+] as const
+
+export type VIEW_CONCRETE_TYPE = (typeof VIEW_CONCRETE_TYPE_VALUES)[number]
+
+export const TABLE_CONCRETE_TYPE_VALUES = [
+  TABLE_ENTITY_CONCRETE_TYPE_VALUE,
+  ...VIEW_CONCRETE_TYPE_VALUES,
+  MATERIALIZED_VIEW_CONCRETE_TYPE_VALUE,
+] as const
+
+export type TABLE_CONCRETE_TYPE = (typeof TABLE_CONCRETE_TYPE_VALUES)[number]
