+++ packages/synapse-react-client/src/lib/utils/synapseTypes/Table/TableEntity.ts
@@ -1,8 +1,5 @@
 import { Table } from './Table'
-
-export const TABLE_ENTITY_CONCRETE_TYPE_VALUE =
-  'org.sagebionetworks.repo.model.table.TableEntity'
-export type TABLE_ENTITY_CONCRETE_TYPE = typeof TABLE_ENTITY_CONCRETE_TYPE_VALUE
+import { TABLE_ENTITY_CONCRETE_TYPE } from './ConcreteType'
 
 // https://docs.synapse.org/rest/org/sagebionetworks/repo/model/table/TableEntity.html
 export interface TableEntity extends Table {
