+++ packages/synapse-react-client/src/lib/utils/synapseTypes/Table/View.ts
@@ -1,19 +1,5 @@
-import {
-  DATASET_CONCRETE_TYPE_VALUE,
-  DATASET_COLLECTION_CONCRETE_TYPE_VALUE,
-  ENTITY_VIEW_CONCRETE_TYPE_VALUE,
-  SUBMISSION_VIEW_CONCRETE_TYPE_VALUE,
-} from '.'
+import { VIEW_CONCRETE_TYPE } from './ConcreteType'
 import { Table } from './Table'
-
-export const VIEW_CONCRETE_TYPE_VALUES = [
-  ENTITY_VIEW_CONCRETE_TYPE_VALUE,
-  SUBMISSION_VIEW_CONCRETE_TYPE_VALUE,
-  DATASET_CONCRETE_TYPE_VALUE,
-  DATASET_COLLECTION_CONCRETE_TYPE_VALUE,
-] as const
-
-export type VIEW_CONCRETE_TYPE = typeof VIEW_CONCRETE_TYPE_VALUES[number]
 
 // https://docs.synapse.org/rest/org/sagebionetworks/repo/model/table/View.html
