+++ packages/synapse-react-client/src/lib/utils/synapseTypes/Table/index.ts
@@ -21,3 +21,4 @@ export * from './Table'
 export * from './TableEntity'
 export * from './TableBundle'
 export * from './View'
+export * from './ConcreteType'
