+++ packages/vite-config/package.json
@@ -6,7 +6,19 @@
   "scripts": {
     "build": "tsc"
   },
-  "exports": "./dist/index.js",
+  "exports": [
+    "./dist/index.js",
+    "./dist/vite-config.js",
+    "./dist/vitest-config.js"
+  ],
+  "peerDependencies": {
+    "vite": "^4.3.4"
+  },
+  "optionalDependencies": {
+    "vitest": "^0.30.1",
+    "@vitest/coverage-c8": "^0.30.1",
+    "@vitest/ui": "^0.30.1"
+  },
   "devDependencies": {
     "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
@@ -15,6 +27,7 @@
     "svgo": "^3.0.2",
     "typescript": "5.0.4",
     "vite": "^4.3.4",
-    "vite-plugin-svgr": "^2.4.0"
+    "vite-plugin-svgr": "^2.4.0",
+    "vitest": "^0.30.1"
   }
 }
